Split up the python code into two files
