+++ Arduino_RPi/.gitignore
@@ -1,1 +1,3 @@
 /output/*
+/fromSD/*
+/processedFiles/*
